+++ src/js/app.js
@@ -12,8 +12,9 @@ import { Hint } from './components/Hint';
 import { actions } from './store';
 import '../css/main.css';
 import { ButtonCountryPTBR } from './components/ButtonPTBR';
-import { ResetButton, clickReset } from './components/ResetButton';
+import { ResetButton } from './components/ResetButton';
 import { ButtonCountryEN } from './components/ButtonEN';
+import { loadButtons, checkResetPTBR } from './appBR';
 
 let resetIdiom = '';
 
@@ -38,17 +39,9 @@ export function setCurrentFromURL(rebus) {
   actions.setCurrent(id);
 }
 
-function checkResetPTBR() {
-  if (localStorage.getItem('flagBR') === 'true') {
-    resetIdiom = 'reset-ptbr';
-  } else {
-    resetIdiom = 'reset-english';
-  }
-}
-
 export function init() {
   // CONDITION: if init as [pt-br] rebus, choose reset idiom
-  checkResetPTBR();
+  resetIdiom = checkResetPTBR(resetIdiom);
 
   try {
     render(
@@ -78,9 +71,7 @@ export function init() {
       ),
       document.querySelector('.root')
     );
-    clickReset(resetIdiom);
-    clickbr();
-    clicken();
+    loadButtons(resetIdiom);
     registerListeners();
   } catch (err) {
     Sentry.captureException(err);
@@ -92,61 +83,4 @@ if (!global || !global.isTestRun) {
   init();
   registerListeners();
   setCurrentFromURL();
-}
-
-// BUTTON [PT-BR] - LOAD BUTTON
-export function clickbr() {
-  // CONDITION: if init as [pt-br] rebus, change design and hide button [ptbr]
-  // ELSE: hide button [english]
-  checkIsPTBR();
-
-  const e = document.getElementById('button-ptbr'); // get button [pt-br]
-
-  // HOVER FUNCTION: cursor (hover) [pt-br]
-  mouseHoverPTBR(e);
-
-  // CLICK FUNCTION: click on button [pt-br]
-  e.addEventListener('click', () => {
-    localStorage.setItem('flagBR', 'true'); // FLAG: now [pt-br] rebus
-  });
-}
-
-// BUTTON [EN] - LOAD BUTTON
-export function clicken() {
-  const e = document.getElementById('button-en'); // get button [english]
-
-  // HOVER FUNCTION: cursor (hover) [english]
-  mouseHoverEN(e);
-
-  // CLICK FUNCTION: click on button [english]
-  e.addEventListener('click', () => {
-    localStorage.setItem('flagBR', 'false'); // FLAG: now [english / original] rebus
-  });
-}
-
-function mouseHoverPTBR(e) {
-  e.addEventListener('mouseover', () => {
-    document.getElementById('button-text-id-br').classList.add('button-text-show');
-  });
-  e.addEventListener('mouseout', () => {
-    document.getElementById('button-text-id-br').classList.remove('button-text-show');
-  });
-}
-
-function mouseHoverEN(e) {
-  e.addEventListener('mouseover', () => {
-    document.getElementById('button-text-id-en').classList.add('button-text-show');
-  });
-  e.addEventListener('mouseout', () => {
-    document.getElementById('button-text-id-en').classList.remove('button-text-show');
-  });
-}
-
-function checkIsPTBR() {
-  if (localStorage.getItem('flagBR') === 'true') {
-    document.querySelector('.root').classList.add('rootbr');
-    document.getElementById('button-ptbr').classList.add('button-spam');
-  } else {
-    document.getElementById('button-en').classList.add('button-spam');
-  }
 }
